Move Wheel class to a separate file
